Refactor: Use final classes to represent instructions
This allows us to remove a lot of hacks in the generator, as all instructions are guaranteed to be reference types. This will also allow us to improve other parts.
All classes are final because the only reason they are classes is because they need to be reference types, they need no inheritance and it is faster that way.